organizing + tweaks
