+++ components/LandingPage/Instructors/data/sampleProblems.ts
