+++ components/LandingPage/Students/data/languages.ts
@@ -1,0 +1,19 @@
+export const languages = [
+  { name: "Python", icon: "devicon-python-plain", origin: "bottom right" },
+  { name: "C++", icon: "devicon-cplusplus-plain", origin: "bottom center" },
+  {
+    name: "JavaScript",
+    icon: "devicon-javascript-plain",
+    origin: "bottom left",
+  },
+  { name: "Java", icon: "devicon-java-plain", origin: "right center" },
+  { name: "Go", icon: "devicon-go-plain", origin: "center center" },
+  { name: "C#", icon: "devicon-csharp-plain", origin: "left center" },
+  { name: "Ruby", icon: "devicon-ruby-plain", origin: "top right" },
+  {
+    name: "TypeScript",
+    icon: "devicon-typescript-plain",
+    origin: "top center",
+  },
+  { name: "Rust", icon: "devicon-rust-plain", origin: "top left" },
+];
